requirements.txt
click==8.0.3
importlib-metadata==4.8.1
numpy==1.21.2
opencv-python==4.5.3.56
python-dotenv==0.19.1
pyzmq==22.3.0
typing-extensions==3.10.0.2
zipp==3.6.0
zmq==0.0.0
